core code updates
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@5669 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/types/IntervalMonthData.java
@@ -105,7 +105,7 @@ public class IntervalMonthData {
         }
     }
 
-    public long getMonths() {
+    public int getMonths() {
         return units;
     }
